pyproject.toml
[tool.poetry]
name = "spiderdjango"
version = "0.1.0"
description = "This is a demonstration of how to use the Django ORM in Spiderweb. This is silly."
authors = ["Joe Kaufeld <opensource@joekaufeld.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
spiderweb-framework = "^1.1.0"
Django = "^5.1.1"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
